+++ src/mjlab/asset_zoo/robots/astribot_s1/__init__.py
@@ -1,0 +1,25 @@
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  ARM_JOINTS as ARM_JOINTS,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  EE_BODY as EE_BODY,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  EE_SITE as EE_SITE,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  HEAD_CAMERA as HEAD_CAMERA,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  HOME_POSE as HOME_POSE,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  TABLE_HEIGHT as TABLE_HEIGHT,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  WORKSPACE_CENTRE as WORKSPACE_CENTRE,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import (
+  get_astribot_s1_cfg as get_astribot_s1_cfg,
+)
+from mjlab.asset_zoo.robots.astribot_s1.s1_constants import get_spec as get_spec